requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==4.6.0